% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{computeUMAP}
\alias{computeUMAP}
\title{UMAP wrapper function}
\usage{
computeUMAP(sce, min_dist = 0.1, n_neighbours = 50, n_dims = 2)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{min_dist}{The effective minimum distance between embedded points.}

\item{n_neighbours}{Integer scalar, number of nearest neighbors to identify
when constructing the initial graph.}

\item{n_dims}{The dimension of the space to embed into.}
}
\value{
SingleCellExperiment object with UMAP results in a reducedDim slot.
}
\description{
Computes UMAP projection on principal components obtained from reducedDim slot.
}
