% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotDimredHighlight}
\alias{plotDimredHighlight}
\title{Plot dimensionality reduction highlighting observations}
\usage{
plotDimredHighlight(
  sce,
  dimred = "TSNE",
  colour_by = "AO_annotation",
  highlight = c("Serotonergic neurons", "Early AP", "Late AP", "Other neural"),
  colours = echino.AO_colours,
  point_size = 1,
  plot_order = NULL,
  legend_position = "right",
  legend_title = T
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{dimred}{A string indicating the reduced dimesnion result in
`reducedDims(sce)` to plot.}

\item{colour_by}{A string indicating the `colData(sce)` column to
colour points by.}

\item{highlight}{A list of strings specifying the `colour_by` values to
highlight.}

\item{colours}{A list of colours of colour palette passed to `scale_color_manual`.}

\item{legend_position}{String passed to legend_position paramater of ggplot theme.}

\item{legend_title}{Boolean indicating whether to plot a legend title.}
}
\description{
Plot dimensionality reduction highlighting observations
}
