% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotUMAP}
\alias{plotUMAP}
\title{Plot UMAP with specific styling}
\usage{
plotUMAP(
  sce,
  colour_by = "celltype",
  point_size = 1,
  legend_position = "right",
  colours,
  alpha = 1
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{colour_by}{The colData observation used to colour points.}

\item{point_size}{Size of the scatter points.}

\item{legend_position}{Position of the colour_by legend. Default is "right".}

\item{celltype_colours}{List of colours to map data values to. The values
will be matched in order.}
}
\description{
Plot UMAP with specific styling
}
